Attempting workflow with yard docs generation.
