+++ Formula/tinker.rb
@@ -1,131 +1,4 @@
-require 'download_strategy'
-
-# Credit to debugging tip:
-#   https://stevenharman.net/debugging-homebrew-with-pry-byebug#a-second-successful-attempt
-def setup_debug_tools
-  Homebrew.install_gem_setup_path! 'pry'
-  Homebrew.install_gem_setup_path! 'pry-byebug', executable: 'pry'
-  Homebrew.install_gem_setup_path! 'dotenv'
-  require 'dotenv/load'
-  require 'pry-byebug'
-end
-
-# Documentation on creating Formulas:
-#   https://docs.brew.sh/Formula-Cookbook
-#
-# Test locally like this:
-#   brew install --debug --verbose --build-from-source Formula/tinker.rb
-
-###
-# Taken from code removed from the homebrew source. They now recommend people
-# define these in their formula. See this link for details:
-#   https://github.com/Homebrew/brew/pull/5112/
-class GitHubPrivateRepositoryDownloadStrategy < CurlDownloadStrategy
-  require 'utils/formatter'
-  require 'utils/github'
-
-  def initialize(url, name, version, **meta)
-    super
-    parse_url_pattern
-    set_github_token
-  end
-
-  def parse_url_pattern
-    unless match = url.match(%r{https://github.com/([^/]+)/([^/]+)(/\S*)*})
-      raise CurlDownloadStrategyError, 'Invalid url pattern for GitHub Repository.'
-    end
-
-    _, @owner, @repo, @filepath = *match
-  end
-
-  def download_url
-    "https://api.github.com/#{@owner}/#{@repo}#{@filepath}"
-  end
-
-  private
-
-  def _fetch(url:, resolved_url:)
-    curl_download(
-      download_url,
-      '--header', "Authorization: token #{@github_token}",
-      to: temporary_path
-    )
-  end
-
-  def set_github_token
-    @github_token = ENV['HOMEBREW_GITHUB_API_TOKEN']
-    unless @github_token
-      raise CurlDownloadStrategyError, 'Environmental variable HOMEBREW_GITHUB_API_TOKEN is required.'
-    end
-
-    validate_github_repository_access!
-  end
-
-  def validate_github_repository_access!
-    # Test access to the repository
-    GitHub.repository(@owner, @repo)
-  rescue GitHub::HTTPNotFoundError
-    # We only handle HTTPNotFoundError here,
-    # becase AuthenticationFailedError is handled within util/github.
-    message = <<~EOS
-      HOMEBREW_GITHUB_API_TOKEN can not access the repository: #{@owner}/#{@repo}
-      This token may not have permission to access the repository or the url of formula may be incorrect.
-    EOS
-    raise CurlDownloadStrategyError, message
-  end
-end
-
-###
-# Taken from code removed from the homebrew source. They now recommend people
-# define these in their formula. See this link for details:
-#   https://github.com/Homebrew/brew/pull/5112/
-class GitHubPrivateRepositoryReleaseDownloadStrategy < GitHubPrivateRepositoryDownloadStrategy
-  def initialize(url, name, version, **meta)
-    super
-  end
-
-  def parse_url_pattern
-    url_pattern = %r{https://github.com/([^/]+)/([^/]+)/archive/([^/]+)(\.tar\.gz|\.zip)}
-    unless @url =~ url_pattern
-      raise CurlDownloadStrategyError, 'Invalid url pattern for GitHub Release.'
-    end
-
-    _, @owner, @repo, @tag, @file_extension = *@url.match(url_pattern)
-    @filename = "#{@tag}#{@file_extension}"
-  end
-
-  def download_url
-    "https://api.github.com/repos/#{@owner}/#{@repo}/releases/assets/#{asset_id}"
-  end
-
-  private
-
-  def _fetch(url:, resolved_url:)
-    curl_download(
-      "#{fetch_release_metadata['tarball_url']}",
-      '--header', 'Accept: application/vnd.github.v3+json',
-      '--header', "Authorization: token #{@github_token}",
-      to: temporary_path
-    )
-  end
-
-  def asset_id
-    @asset_id ||= resolve_asset_id
-  end
-
-  def resolve_asset_id
-    release_metadata = fetch_release_metadata
-    assets = release_metadata['assets'].select { |a| a['name'] == @filename }
-    raise CurlDownloadStrategyError, 'Asset file not found.' if assets.empty?
-
-    assets.first['id']
-  end
-
-  def fetch_release_metadata
-    release_url = "https://api.github.com/repos/#{@owner}/#{@repo}/releases/tags/#{@tag}"
-    GitHub.open_api(release_url)
-  end
-end
+require_relative '../lib/git_hub_private_repository_release_download_strategy'
 
 class Tinker < Formula
   desc 'Install the Tinker toolset.'
